PatikaStore
PatikaStore

Now, our route will produce goods. You must contribute to the design as well.

As shown in the design, you must create a search bar and a product listing. For the time being, the search bar's features are not required. You must make changes to the React Native document for the component where user text input is possible.

FlatList must be used for listing. A two-column display similar to the one on the screen has some documentation available for you to peruse.

You can also get the JSON data you'll need below.

Convenience:)